arm64: dts: Tweak QG SOC to be more linear/monotonic This enables some tweaks to make QGauge's SSOC and MSOC values more linear and monotonic, since the UI SOC curve currently drops very slowly at the beginning (i.e. near 100%) and accelerates later. A more linear SOC curve is helpful for battery testing and improves UX.
This enables the following QG options as seen in the SM7250 QRD:
  - Hold SOC while full: show 100% once full, prerequisite for linearize
    Recharging is normally done between 95-100%, so the base MSOC may
    fluctuate otherwise
  - Linearize SOC: scale MSOC to be more linear after 100%
  - TCSS: scale SSOC with charge termination current after 90% to avoid
    jumps during charging
  - BASS: scale SSOC with BSOC to mitigate jumps

Note that this only adjusts the QG source SOCs; further adjustments to
GBMS may be needed to fully linearize the UI SOC curve as it derives the
final SSOC value from QG CC SOC, raw SOC, MSOC, BSOC, SSOC, delta SOC,
and other math and values from the DT config. For now, this should still
help slightly and we can compare it with GBMS SOC data to decide what to
do about the GBMS side of UI SOC calculation later.

Signed-off-by: onettboots <blackcocopet@gmail.com>